Make ClusteredTopDocsExternalizer an inner class of NodeTopDocs
