+++ query/src/main/java/org/infinispan/query/clustered/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * @private
+ */
+package org.infinispan.query.clustered;
