+++ query/src/main/java/org/infinispan/query/impl/externalizers/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Externalizers for several Lucene objects.
+ *
+ * @private
+ */
+package org.infinispan.query.impl.externalizers;
